requirements.txt
torch>=2.1
torchvision>=0.16
PyYAML>=6.0
Pillow>=10.0
tqdm>=4.66
numpy>=1.26
